+++ lib/granalytics.rb
@@ -17,6 +17,7 @@ require_dependency 'granalytics/configuration'
 require_dependency 'granalytics/event'
 require_dependency 'granalytics/aggregate'
 require_dependency 'granalytics/data'
+require_dependency 'granalytics/export'
 
 if ::Rails
   require "granalytics/rails.rb"
